requirements.txt
click==8.1.7
dataloader==2.0
efficientnet==1.0.0
einops==0.6.1
image-classifiers==1.0.0
imageio==2.27.0
keras==2.9.0
numpy==1.24.4
packaging==23.1
pandas==2.0.3
Pillow==10.0.1
promise==2.3
protobuf==3.20.3
psutil==5.9.5
python-dateutil==2.8.2
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
segmentation-models==1.0.1
sklearn==0.0
sympy==1.12
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-datasets==4.9.2
tensorflow-estimator==2.9.0
tensorflow-examples @ git+https://github.com/tensorflow/examples.git@2741d0f2a345ef52e0de1d498de732f5dacbf2f0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-metadata==1.14.0
torch==2.0.1
torchmetrics==1.1.2
torchview==0.2.6
torchvision==0.15.2
tqdm==4.66.1

